Added Support For Displaying Routes To Select
I have added a good bit of code to support the displaying of multiple routes for the user to select. Once selected the app should show the selected route on the Route display screen. However I was unable to test it as we are having trouble pulling objects down from the database. Routes can be uploaded but not downloaded.

-MacKenzie